refactor: move Polars dataframe serializer to new file
